+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+* **New Features**
+  * Support list/map nearest key/value get/remove operations (relative rank range). Requires server version v4.3.0 or later. [#264](https://github.com/aerospike/aerospike-client-nodejs/pull/264)
+  * Support list write flags NO_FAIL and PARTIAL. Add new map write flags, including NO_FAIL and PARTIAL. Requires server version v4.3.0 or later. [#265](https://github.com/aerospike/aerospike-client-nodejs/pull/265)
+
 * **Updates**
   * Update C client library to [v4.3.14](http://www.aerospike.com/download/client/c/notes.html#4.3.14) [#263](https://github.com/aerospike/aerospike-client-nodejs/pull/263)